Cherry pick changes prior to R update.

+++ Models/HMM/Clickstream/Event.cpp
@@ -18,16 +18,15 @@
 */
 
 #include "Models/HMM/Clickstream/Event.hpp"
-#include "cpputil/math_utils.hpp"
 
 namespace BOOM {
   namespace Clickstream {
 
     Event::Event(int event_type, const Ptr<CatKeyBase> &key)
-        : MarkovData(event_type, key) {}
+        : MarkovData(uint(event_type), key) {}
 
     Event::Event(int event_type, const Ptr<Event> &prev)
-        : MarkovData(event_type, Ptr<MarkovData>(prev)) {}
+        : MarkovData(uint(event_type), Ptr<MarkovData>(prev)) {}
 
     Event *Event::clone() const { return new Event(*this); }
